Fix compile error under 64-bit Rust

+++ src/state.rs
@@ -1365,7 +1365,7 @@ impl State {
   /// Maps to `luaL_loadbuffer`.
   pub fn load_buffer(&mut self, buff: &[u8], name: &str) -> ThreadStatus {
     let name_c_str = CString::new(name).unwrap();
-    let result = unsafe { ffi::luaL_loadbuffer(self.L, buff.as_ptr() as *const _, buff.len() as u32, name_c_str.as_ptr()) };
+    let result = unsafe { ffi::luaL_loadbuffer(self.L, buff.as_ptr() as *const _, buff.len() as size_t, name_c_str.as_ptr()) };
     ThreadStatus::from_c_int(result).unwrap()
   }
